Fully integrate support for LVS collision constraints
