+++ trajopt_ifopt/include/trajopt_ifopt/trajopt_ifopt.h
@@ -24,8 +24,14 @@
 #ifndef TRAJOPT_IFOPT_TRAJOPT_IFOPT_H
 #define TRAJOPT_IFOPT_TRAJOPT_IFOPT_H
 
+#include <trajopt_ifopt/constraints/discrete_collision_constraint.h>
+#include <trajopt_ifopt/constraints/lvs_collision_constraint.h>
+#include <trajopt_ifopt/constraints/discrete_collision_evaluators.h>
+#include <trajopt_ifopt/constraints/lvs_collision_evaluators.h>
+#include <trajopt_ifopt/constraints/collision_utils.h>
+#include <trajopt_ifopt/constraints/collision_types.h>
+
 #include <trajopt_ifopt/constraints/cartesian_position_constraint.h>
-#include <trajopt_ifopt/constraints/collision_constraint.h>
 #include <trajopt_ifopt/constraints/inverse_kinematics_constraint.h>
 #include <trajopt_ifopt/constraints/joint_position_constraint.h>
 #include <trajopt_ifopt/constraints/joint_velocity_constraint.h>
