+++ trajopt_optimizers/trajopt_sqp/include/trajopt_sqp/callbacks/collision_plotter.h
@@ -32,7 +32,7 @@ TRAJOPT_IGNORE_WARNINGS_PUSH
 #include <tesseract_visualization/visualization.h>
 TRAJOPT_IGNORE_WARNINGS_POP
 
-#include <trajopt_ifopt/constraints/collision_constraint.h>
+#include <trajopt_ifopt/constraints/discrete_collision_constraint.h>
 #include <trajopt_sqp/sqp_callback.h>
 #include <trajopt_sqp/trust_region_sqp_solver.h>
 #include <trajopt_sqp/types.h>
@@ -49,14 +49,14 @@ public:
 
   void plot(const ifopt::Problem& nlp);
 
-  void addConstraintSet(const trajopt::CollisionConstraintIfopt::ConstPtr& collision_constraint);
+  void addConstraintSet(const trajopt::DiscreteCollisionConstraintIfopt::ConstPtr& collision_constraint);
 
-  void addConstraintSet(const std::vector<trajopt::CollisionConstraintIfopt::ConstPtr>& collision_constraints);
+  void addConstraintSet(const std::vector<trajopt::DiscreteCollisionConstraintIfopt::ConstPtr>& collision_constraints);
 
   bool execute(const ifopt::Problem& nlp, const trajopt_sqp::SQPResults& sqp_results) override;
 
 protected:
-  std::vector<trajopt::CollisionConstraintIfopt::ConstPtr> collision_constraints_;
+  std::vector<trajopt::DiscreteCollisionConstraintIfopt::ConstPtr> collision_constraints_;
   tesseract_visualization::Visualization::Ptr plotter_;
 };
 }  // namespace trajopt_sqp
